add authentication to CRUED operations

+++ pages/api/products/[id].js
@@ -2,7 +2,9 @@ import dbConnect from "../../../util/mongo";
 import Product from "../../../models/Product";
 
 export default async function handler(req, res) {
-    const { method, query: { id } } = req;  
+    const { method, query: { id }, cookies } = req;  
+
+    const token = cookies.token;
 
     dbConnect(); 
 
@@ -16,6 +18,9 @@ export default async function handler(req, res) {
     }
 
     if (method === "PUT") {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json("Not authenticated! ");
+        }
         try {
             const product = await Product.findByIdAndUpdate(id, req.body, {new: true});
             res.status(200).json(product);
@@ -25,6 +30,9 @@ export default async function handler(req, res) {
     }
 
     if (method === "DELETE") {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json("Not authenticated! ");
+        }
         try {
             await Product.findByIdAndDelete(id);
             res.status(200).json("The product has been deleted!");
